Deletes commented-out sections and moves utility functions from FourierFlows.jl to utils.jl
